mapper

+++ User1/json_req.h
@@ -14,8 +14,8 @@ namespace req_json {
     {
         std::string file_name;
         int oper_num;
-        int start;
-        int end;
+        long start;
+        long end;
     };
 
     void tag_invoke(value_from_tag, value& jv, req const& c)
@@ -34,8 +34,8 @@ namespace req_json {
         return req{
             value_to<std::string>(obj.at("file_name")),
             value_to<int>(obj.at("oper_num")),
-            value_to<int>(obj.at("start")),
-            value_to<int>(obj.at("end"))
+            value_to<long>(obj.at("start")),
+            value_to<long>(obj.at("end"))
         };
     }
